rajouté des trucs

+++ ft_strcpy.c
@@ -6,13 +6,13 @@
 /*   By: fdubois <fdubois@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/08/10 16:03:41 by fdubois           #+#    #+#             */
-/*   Updated: 2018/08/10 16:20:37 by fdubois          ###   ########.fr       */
+/*   Updated: 2018/08/12 16:44:25 by fdubois          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 char *ft_strcpy(char *dest, char *src)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (src[i] != '\0')
